🧪 Populate tests/test_security.py with comprehensive SSRF tests
